requirements.txt
rasterio
shapely
geopandas
numpy
tqdm
mercantile
Pillow